requirements.txt
Flask
tensorflow
keras
pillow
numpy
pandas
opencv-python
gunicorn
